Extract walkApi
