+++ createClient.js
@@ -2,6 +2,7 @@ var muxrpc = require('muxrpc')
 var setIn = require('set-in')
 
 var defaultSerialize = require('./serialize')
+var walkApi = require('./walkApi')
 
 module.exports = createClient
 
@@ -18,14 +19,3 @@ function createClient (api, config) {
   return muxrpc(manifest, null, serialize)()
 }
 
-function walkApi (api, cb, path) {
-  path = path || []
-  Object.keys(api).forEach(function (name) {
-    var service = api[name]
-    var servicePath = path.concat([name])
-    cb(service, servicePath)
-    if (service.api) {
-      walkApi(api, cb, servicePath)
-    }
-  })
-}
